@@ -1,5 +1,5 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' as mat;
 import 'package:one_context/src/controllers/one_context.dart';
 
 typedef Widget DialogBuilder(BuildContext context);
@@ -7,8 +7,17 @@ mixin DialogController {
   /// Return dialog utility class `DialogController`
   DialogController get dialog => this;
 
-  /// The current context
-  BuildContext? get context => OneContext().context;
+  /// The current scaffold state
+  /// It is used to show snackbars, dialogs, modalBottomsheets, etc.
+  GlobalKey<ScaffoldState>? _scaffoldKey;
+  GlobalKey<ScaffoldState> get scaffoldKey {
+    _scaffoldKey ??= GlobalKey<ScaffoldState>();
+    return _scaffoldKey!;
+  }
+
+  set scaffoldKey(scaffKey) => _scaffoldKey = scaffKey;
+  BuildContext? get _scaffoldContext => _scaffoldKey?.currentContext;
+  ScaffoldState? get _scaffoldState => _scaffoldKey?.currentState;
 
   ValueNotifier<List<Widget>> _dialogs = ValueNotifier([]);
   ValueNotifier<List<Widget>> get dialogNotifier => _dialogs;
@@ -18,6 +27,7 @@ mixin DialogController {
     _dialogs.value.add(widget);
   }
 
+  /// Removes the last dialog
   void removeDialogVisible({Widget? widget}) {
     if (widget != null) {
       _dialogs.value.remove(widget);
@@ -25,6 +35,7 @@ mixin DialogController {
       _dialogs.value.removeLast();
   }
 
+  /// Pop all dialogs
   void popAllDialogs() {
     _dialogs.value.forEach((element) {
       OneContext().popDialog();
@@ -36,48 +47,6 @@ mixin DialogController {
     _dialogs.value.clear();
   }
 
-  Future<T?> Function<T>({
-    required Widget Function(BuildContext) builder,
-    bool? barrierDismissible,
-    bool useRootNavigator,
-    Color? barrierColor,
-    String? barrierLabel,
-    bool useSafeArea,
-    RouteSettings? routeSettings,
-    Offset? anchorPoint,
-  })? _showDialog;
-
-  Future<T?> Function<T>({
-    required Widget Function(BuildContext) builder,
-    Color? backgroundColor,
-    double? elevation,
-    ShapeBorder? shape,
-    Clip? clipBehavior,
-    bool isScrollControlled,
-    bool useRootNavigator,
-    bool isDismissible,
-    BoxConstraints? constraints,
-    Color? barrierColor,
-    bool? enableDrag,
-    RouteSettings? routeSettings,
-    AnimationController? transitionAnimationController,
-    Offset? anchorPoint,
-  })? _showModalBottomSheet;
-
-  ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-      SnackBar Function(BuildContext?) builder)? _showSnackBar;
-
-  PersistentBottomSheetController<T> Function<T>({
-    Widget Function(BuildContext) builder,
-    Color? backgroundColor,
-    double? elevation,
-    ShapeBorder? shape,
-    Clip? clipBehavior,
-    BoxConstraints? constraints,
-    bool? enableDrag,
-    AnimationController? transitionAnimationController,
-  })? _showBottomSheet;
-
   /// Displays a Material dialog above the current contents of the app, with
   /// Material entrance and exit animations, modal barrier color, and modal
   /// barrier behavior (dialog is dismissible with a tap on the barrier).
@@ -87,27 +56,30 @@ mixin DialogController {
     String? barrierLabel,
     RouteSettings? routeSettings,
     Offset? anchorPoint,
-    bool barrierDismissible = true,
+    bool? barrierDismissible = true,
     Color? barrierColor = Colors.black54,
     bool useSafeArea = true,
+    TraversalEdgeBehavior? traversalEdgeBehavior,
   }) async {
-    if (!(await _contextLoaded())) return null;
+    if (!(await _scaffoldContextLoaded())) return null;
 
-    Widget dialog = builder(context!);
-    if (barrierDismissible == true) addDialogVisible(dialog);
+    Widget dialog = builder(_scaffoldContext!);
+    addDialogVisible(dialog);
 
-    return _showDialog!<T>(
-      builder: (_) => dialog,
-      barrierDismissible: barrierDismissible,
-      useRootNavigator: useRootNavigator,
-      barrierColor: barrierColor,
-      barrierLabel: barrierLabel,
-      useSafeArea: useSafeArea,
-      routeSettings: routeSettings,
-      anchorPoint: anchorPoint,
-    ).whenComplete(() {
-      if (barrierDismissible == true) removeDialogVisible(widget: dialog);
-    });
+    return mat
+        .showDialog<T>(
+          context: _scaffoldContext!,
+          builder: (_) => dialog,
+          barrierDismissible: barrierDismissible ?? true,
+          useRootNavigator: useRootNavigator,
+          barrierColor: barrierColor,
+          barrierLabel: barrierLabel,
+          useSafeArea: useSafeArea,
+          routeSettings: routeSettings,
+          anchorPoint: anchorPoint,
+          traversalEdgeBehavior: traversalEdgeBehavior,
+        )
+        .whenComplete(() => removeDialogVisible(widget: dialog));
   }
 
   /// Removes the current [SnackBar] by running its normal exit animation.
@@ -115,8 +87,8 @@ mixin DialogController {
   /// The closed completer is called after the animation is complete.
   void hideCurrentSnackBar(
       {SnackBarClosedReason reason = SnackBarClosedReason.hide}) async {
-    if (!(await _contextLoaded())) return;
-    ScaffoldMessenger.of(context!).hideCurrentSnackBar(reason: reason);
+    if (!(await _scaffoldContextLoaded())) return;
+    ScaffoldMessenger.of(_scaffoldContext!).hideCurrentSnackBar(reason: reason);
   }
 
   /// Removes the current [SnackBar] (if any) immediately.
@@ -125,15 +97,17 @@ mixin DialogController {
   /// any queued snack bars, they begin their entrance animation immediately.
   void removeCurrentSnackBar(
       {SnackBarClosedReason reason = SnackBarClosedReason.hide}) async {
-    if (!(await _contextLoaded())) return;
-    ScaffoldMessenger.of(context!).removeCurrentSnackBar(reason: reason);
+    if (!(await _scaffoldContextLoaded())) return;
+    ScaffoldMessenger.of(_scaffoldContext!)
+        .removeCurrentSnackBar(reason: reason);
   }
 
   /// Shows a [SnackBar] at the bottom of the scaffold.
   Future<ScaffoldFeatureController<SnackBar, SnackBarClosedReason>?>
       showSnackBar({required SnackBar Function(BuildContext?) builder}) async {
-    if (!(await _contextLoaded())) return null;
-    return _showSnackBar!(builder);
+    if (!(await _scaffoldContextLoaded())) return null;
+    return ScaffoldMessenger.of(_scaffoldContext!)
+        .showSnackBar(builder(_scaffoldContext));
   }
 
   /// Shows a modal material design bottom sheet.
@@ -155,37 +129,38 @@ mixin DialogController {
     RouteSettings? routeSettings,
     AnimationController? transitionAnimationController,
     Offset? anchorPoint,
+    bool? showDragHandle,
+    bool useSafeArea = false,
   }) async {
-    if (!(await _contextLoaded())) return null;
+    if (!(await _scaffoldContextLoaded())) return null;
 
-    Widget dialog = builder(context!);
-    if (isDismissible == true) addDialogVisible(dialog);
+    Widget dialog = builder(_scaffoldContext!);
+    addDialogVisible(dialog);
 
-    return _showModalBottomSheet!<T>(
-      builder: builder,
-      backgroundColor: backgroundColor,
-      clipBehavior: clipBehavior,
-      elevation: elevation,
-      isDismissible: isDismissible,
-      isScrollControlled: isScrollControlled,
-      shape: shape,
-      useRootNavigator: useRootNavigator,
-      constraints: constraints,
-      barrierColor: barrierColor,
-      enableDrag: enableDrag,
-      routeSettings: routeSettings,
-      transitionAnimationController: transitionAnimationController,
-      anchorPoint: anchorPoint,
-    ).whenComplete(() {
-      if (isDismissible == true) removeDialogVisible(widget: dialog);
-    });
+    return mat
+        .showModalBottomSheet<T>(
+          context: _scaffoldContext!,
+          builder: (_) => dialog,
+          backgroundColor: backgroundColor,
+          clipBehavior: clipBehavior,
+          elevation: elevation,
+          isDismissible: isDismissible,
+          isScrollControlled: isScrollControlled,
+          shape: shape,
+          useRootNavigator: useRootNavigator,
+          constraints: constraints,
+          barrierColor: barrierColor,
+          enableDrag: enableDrag = true,
+          routeSettings: routeSettings,
+          transitionAnimationController: transitionAnimationController,
+          anchorPoint: anchorPoint,
+          showDragHandle: showDragHandle,
+          useSafeArea: useSafeArea,
+        )
+        .whenComplete(() => removeDialogVisible(widget: dialog));
   }
 
-  /// Shows a material design bottom sheet in the nearest [Scaffold] ancestor. If
-  /// you wish to show a persistent bottom sheet, use [Scaffold.bottomSheet].
-  ///
-  /// Returns a controller that can be used to close and otherwise manipulate the
-  /// bottom sheet.
+  /// Shows a persistent bottom sheet
   Future<PersistentBottomSheetController<T>?> showBottomSheet<T>({
     required Widget Function(BuildContext) builder,
     Color? backgroundColor,
@@ -196,9 +171,13 @@ mixin DialogController {
     bool? enableDrag,
     AnimationController? transitionAnimationController,
   }) async {
-    if (!(await _contextLoaded())) return null;
-    return _showBottomSheet!<T>(
-      builder: builder,
+    if (!(await _scaffoldContextLoaded())) return null;
+
+    Widget dialog = builder(_scaffoldContext!);
+    addDialogVisible(dialog);
+
+    return _scaffoldState!.showBottomSheet(
+      (_) => dialog,
       backgroundColor: backgroundColor,
       elevation: elevation,
       shape: shape,
@@ -206,70 +185,27 @@ mixin DialogController {
       constraints: constraints,
       enableDrag: enableDrag,
       transitionAnimationController: transitionAnimationController,
-    );
-  }
-
-  /// Register callbacks
-  void registerCallback(
-      {Future<T?> Function<T>({
-        required Widget Function(BuildContext) builder,
-        bool? barrierDismissible,
-        bool useRootNavigator,
-        Color? barrierColor,
-        String? barrierLabel,
-        bool useSafeArea,
-        RouteSettings? routeSettings,
-        Offset? anchorPoint,
-      })?
-          showDialog,
-      Future<T?> Function<T>({
-        required Widget Function(BuildContext) builder,
-        Color? backgroundColor,
-        double? elevation,
-        ShapeBorder? shape,
-        Clip? clipBehavior,
-        bool? isScrollControlled,
-        bool? useRootNavigator,
-        bool? isDismissible,
-        BoxConstraints? constraints,
-        Color? barrierColor,
-        bool? enableDrag,
-        RouteSettings? routeSettings,
-        AnimationController? transitionAnimationController,
-        Offset? anchorPoint,
-      })?
-          showModalBottomSheet,
-      ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-              SnackBar Function(BuildContext?) builder)?
-          showSnackBar,
-      PersistentBottomSheetController<T> Function<T>({
-        Widget Function(BuildContext)? builder,
-        Color? backgroundColor,
-        double? elevation,
-        ShapeBorder? shape,
-        Clip? clipBehavior,
-        BoxConstraints? constraints,
-        bool? enableDrag,
-        AnimationController? transitionAnimationController,
-      })?
-          showBottomSheet}) {
-    _showDialog = showDialog;
-    _showSnackBar = showSnackBar;
-    _showModalBottomSheet = showModalBottomSheet;
-    _showBottomSheet = showBottomSheet;
+    )..closed.whenComplete(() => removeDialogVisible(widget: dialog));
   }
 
   /// Pop the top-most dialog off the OneContext.dialog.
   popDialog<T extends Object>([T? result]) async {
-    if (!(await _contextLoaded())) return;
-    return Navigator.of(context!).pop<T>(result);
+    if ((await _scaffoldContextLoaded())) {
+      if (OneContext().hasDialogVisible)
+        return Navigator.of(_scaffoldContext!).pop<T>(result);
+    }
   }
 
-  Future<bool> _contextLoaded() async {
+  /// Context used by inner Navigator
+  Future<bool> _scaffoldContextLoaded() async {
     await Future.delayed(Duration.zero);
-    if (!OneContext.hasContext && !kReleaseMode) {
+    final isContextNull = _scaffoldContext == null;
+    final isMounted = _scaffoldContext?.mounted ?? false;
+
+    if ((isContextNull || !isMounted)) {
       throw NO_CONTEXT_ERROR;
     }
-    return OneContext.hasContext;
+
+    return !isContextNull;
   }
 }
