@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:one_context/src/components/one_context_widget.dart';
 import 'package:one_context/src/controllers/dialog_controller.mixin.dart';
 import 'package:one_context/src/controllers/navigator_controller.mixin.dart';
 import 'package:one_context/src/controllers/one_notification_controller.dart';
-import 'package:one_context/src/controllers/one_theme_controller.dart';
 import 'package:one_context/src/controllers/overlay_controller.mixin.dart';
+
+import '../../one_context.dart';
 
 class OneContext with NavigatorController, OverlayController, DialogController {
   static BuildContext? _context;
@@ -12,11 +12,13 @@ class OneContext with NavigatorController, OverlayController, DialogController {
   /// The almost top root context of the app,
   /// use it carefully or don't use it directly!
   BuildContext? get context {
+    _context = key.currentContext;
     assert(_context != null, NO_CONTEXT_ERROR);
     return _context;
   }
 
-  static bool get hasContext => _context != null;
+  static bool get hasContext => OneContext().context != null;
+
   set context(BuildContext? newContext) => _context = newContext;
 
   /// If you need reactive changes, do not use OneContext().mediaQuery
@@ -56,58 +58,6 @@ class OneContext with NavigatorController, OverlayController, DialogController {
   static OneContext instance = OneContext._private();
   factory OneContext() => instance;
 
-  /// Register all necessary callbacks from main widget, automatically
-  void registerDialogCallback({
-    Future<T?> Function<T>({
-      required Widget Function(BuildContext) builder,
-      bool? barrierDismissible,
-      bool useRootNavigator,
-      Color? barrierColor,
-      String? barrierLabel,
-      bool useSafeArea,
-      RouteSettings? routeSettings,
-      Offset? anchorPoint,
-    })?
-        showDialog,
-    Future<T?> Function<T>({
-      required Widget Function(BuildContext) builder,
-      Color? backgroundColor,
-      double? elevation,
-      ShapeBorder? shape,
-      Clip? clipBehavior,
-      bool? isScrollControlled,
-      bool? useRootNavigator,
-      bool? isDismissible,
-      BoxConstraints? constraints,
-      Color? barrierColor,
-      bool? enableDrag,
-      RouteSettings? routeSettings,
-      AnimationController? transitionAnimationController,
-      Offset? anchorPoint,
-    })?
-        showModalBottomSheet,
-    ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-            SnackBar Function(BuildContext?) builder)?
-        showSnackBar,
-    PersistentBottomSheetController<T> Function<T>({
-      Widget Function(BuildContext)? builder,
-      Color? backgroundColor,
-      double? elevation,
-      ShapeBorder? shape,
-      Clip? clipBehavior,
-      BoxConstraints? constraints,
-      bool? enableDrag,
-      AnimationController? transitionAnimationController,
-    })?
-        showBottomSheet,
-  }) {
-    registerCallback(
-        showDialog: showDialog,
-        showSnackBar: showSnackBar,
-        showModalBottomSheet: showModalBottomSheet,
-        showBottomSheet: showBottomSheet);
-  }
-
   /// Use [OneContext().builder] in MaterialApp builder,
   /// in order to show dialogs and overlays.
   ///
@@ -118,57 +68,20 @@ class OneContext with NavigatorController, OverlayController, DialogController {
   ///       builder: OneContext().builder,
   ///      ...
   /// ```
-  Widget builder(BuildContext context, Widget? widget,
-          {Key? key,
-          MediaQueryData? mediaQueryData,
-          String? initialRoute,
-          Route<dynamic> Function(RouteSettings)? onGenerateRoute,
-          Route<dynamic> Function(RouteSettings)? onUnknownRoute,
-          List<NavigatorObserver> observers = const <NavigatorObserver>[]}) =>
-      ParentContextWidget(
+  Widget builder(
+    BuildContext context,
+    Widget? widget, {
+    Key? key,
+    MediaQueryData? mediaQueryData,
+    String? initialRoute,
+    List<NavigatorObserver> observers = const <NavigatorObserver>[],
+  }) =>
+      OneContextWidget(
         child: widget,
         mediaQueryData: mediaQueryData,
         initialRoute: initialRoute,
-        onGenerateRoute: onGenerateRoute,
-        onUnknownRoute: onUnknownRoute,
         observers: observers,
       );
-}
-
-class ParentContextWidget extends StatelessWidget {
-  final MediaQueryData? mediaQueryData;
-  final String? initialRoute;
-  final Route<dynamic> Function(RouteSettings)? onGenerateRoute;
-  final Route<dynamic> Function(RouteSettings)? onUnknownRoute;
-  final List<NavigatorObserver> observers;
-  final Widget? child;
-
-  const ParentContextWidget(
-      {Key? key,
-      this.child,
-      this.mediaQueryData,
-      this.initialRoute,
-      this.onGenerateRoute,
-      this.onUnknownRoute,
-      this.observers = const <NavigatorObserver>[]})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MediaQuery(
-      data: mediaQueryData ?? MediaQuery.of(context),
-      child: Navigator(
-        initialRoute: initialRoute,
-        onUnknownRoute: onUnknownRoute,
-        observers: observers,
-        onGenerateRoute: onGenerateRoute ??
-            (settings) => MaterialPageRoute(
-                builder: (context) => OneContextWidget(
-                      child: child,
-                    )),
-      ),
-    );
-  }
 }
 
 const String NO_CONTEXT_ERROR = """
